omfg windows. not yet tho

+++ sources/main/random.cpp
@@ -10,7 +10,7 @@
 // Private
 
 int rand(int max) {
-	static std::mt19937 prng{std::time(nullptr)};
+	static std::mt19937 prng{static_cast<unsigned int>(std::time(nullptr))};
 	std::uniform_int_distribution<> dist{0, max - 1};
 
 	return dist(prng);
